fix(overlay): updated file structure, moved all ui components in ui folder for spotlight design system
