+++ packages/overlay/src/integrations/sentry/components/events/EventList.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useSpotlightContext } from '~/lib/useSpotlightContext';
-import Badge from '../../../../components/Badge';
+import Badge from '~/ui/Badge';
 import CardList from '../../../../components/CardList';
 import TimeSince from '../../../../components/TimeSince';
 import { useSentryEvents } from '../../data/useSentryEvents';
