+++ packages/overlay/src/integrations/sentry/components/performance/QuerySummary.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import { ReactComponent as Sort } from '~/assets/sort.svg';
 import { ReactComponent as SortDown } from '~/assets/sortDown.svg';
-import Breadcrumbs from '~/components/Breadcrumbs';
 import classNames from '~/lib/classNames';
+import Breadcrumbs from '~/ui/Breadcrumbs';
 import { QUERY_SUMMARY_HEADERS, QUERY_SUMMARY_SORT_KEYS } from '../../constants';
 import { useSentrySpans } from '../../data/useSentrySpans';
 import { Span } from '../../types';
