+++ packages/overlay/src/ui/Badge/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Badge';
