+++ packages/overlay/src/ui/Breadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Breadcrumbs';
