+++ packages/overlay/src/ui/Chart/RingChart/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RingChart';
