+++ packages/overlay/src/ui/Chart/index.ts
@@ -1,0 +1,1 @@
+export { default as RingChart } from './RingChart';
