+++ packages/overlay/src/ui/SidePanel/index.ts
@@ -1,0 +1,1 @@
+export { SidePanelHeader, default } from './SidePanel';
