+++ packages/overlay/src/ui/Tag/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Tag';
